Reuse Mock user secrets manager
